DRAFT: Add in Subscription dataclass
This adds in the `Subscription` dataclass, which stores state about a subscription and implements `build_request`, which returns a `(Dataset, Request)` tuple that can be used to run a query.

Currently this is tested using `parse_and_run_query`. This works fine at the moment as a test, but in production we don't want subscription queries to interfere with api queries, and vice versa. I don't know whether we want to rate limit subscription queries at all at this point, but we definitely want them to be using a different set of buckets to api queries if they do.

So we have two main directions we could take this:
 - We could refactor `parse_and_run_query` so that we can run a query and just skip rate limiting, max_threads, etc.
 - We could override the request settings to for subscription queries so that we return no rate limits, then modify the ratelimiter to skip all the calls to redis if no rate limits are set.

Don't pay too much attention to the test, it's just there to make sure running the query runs at all. It's mostly copied from other tests, I'll factor that better once we decide how to move forward.